refactor(terminal): keep reattach mouse modes by not resetting them
Replaces the renderer-side replay rescan with a reset profile that omits
RESET_MOUSE_REPORTING on an alternate-screen reattach. The daemon's
rehydrateSequences already re-arm the live TUI's protocol and encoding inside
the replayed payload, so xterm's own parse is the authority — the renderer only
had to stop clobbering it. Alt-screen comes from the daemon's authoritative
flag, falling back to the kitty tracker that already scanned the same payload.

Drops the shared DECSET rescan module (196 lines) and its two full-payload
regex passes per reattach; the payload is uncapped up to the 16MB daemon frame
and this ran synchronously on the renderer main thread. The daemon mouse-mode
mirror stays in the daemon, where the only full emulator lives.
